Alteração da estrutura do projeto para se adequar a estrutura de Model View Controller
