+++ .gitignore
@@ -1,1 +1,26 @@
 config_URI.py
+__pycache__/auth_utils.cpython-310.pyc
+__pycache__/config_URI.cpython-310.pyc
+__pycache__/database.cpython-310.pyc
+__pycache__/logging_config.cpython-310.pyc
+__pycache__/main.cpython-310.pyc
+__pycache__/models.cpython-310.pyc
+__pycache__/config_URI.cpython-310.pyc
+__pycache__/main.cpython-310.pyc
+__pycache__/config_URI.cpython-310.pyc
+__pycache__/main.cpython-310.pyc
+controller/__pycache__/__init__.cpython-310.pyc
+controller/__pycache__/auth_utils.cpython-310.pyc
+controller/__pycache__/config_URI.cpython-310.pyc
+controller/__pycache__/database.cpython-310.pyc
+controller/__pycache__/main.cpython-310.pyc
+logs/__pycache__/logging_config.cpython-310.pyc
+model/__pycache__/models.cpython-310.pyc
+testes/__pycache__/__init__.cpython-310.pyc
+testes/__pycache__/test_main.cpython-310-pytest-8.3.5.pyc
+.gitignore
+__pycache__/config_URI.cpython-310.pyc
+__pycache__/main.cpython-310.pyc
+view/__pycache__/__init__.cpython-310.pyc
+view/__pycache__/tasks.cpython-310.pyc
+view/__pycache__/users.cpython-310.pyc
