--- estrutura.txt
@@ -1,25 +1,0 @@
-task_manager/
-├── manage.py
-├── task_manager/   # Diretório principal do projeto Django.
-│   ├── __init__.py
-│   ├── settings.py
-│   ├── urls.py
-│   ├── wsgi.py
-├── core/   # App Django onde vai ficar a lógica da API (usuários, tarefas, autenticação).
-│   ├── __init__.py
-│   ├── models.py         # Modelos (usuários e tarefas)
-│   ├── serializers.py    # Serializadores para converter objetos em JSON
-│   ├── views.py         # Views da API (endpoints)
-│   ├── urls.py          # Rotas específicas do app
-│   ├── tests/  # Pasta para testes automatizados, organizada por funcionalidade.
-│   │   ├── __init__.py
-│   │   ├── test_users.py
-│   │   ├── test_tasks.py
-│   │   ├── test_auth.py
-├── docs/   # Pasta para documentação e diagramas.
-│   ├── README.md        # Documentação principal
-│   ├── diagrams/
-│   │   ├── data_model.mmd  # Diagrama de dados (Mermaid)
-│   │   ├── request_flow.mmd # Fluxo de requisições (Mermaid)
-├── requirements.txt      # Dependências
-├── .gitignore           # Arquivos ignorados pelo Git
