+++ logs/logging_config.py
@@ -1,0 +1,8 @@
+import logging
+
+def setup_logging():
+    logging.basicConfig(
+        filename='logs/app.log',
+        level=logging.INFO,
+        format='%(asctime)s %(levelname)s:%(message)s'
+    )
