+++ model/models.py
@@ -1,0 +1,44 @@
+from pydantic import BaseModel
+from typing import Optional
+
+class User(BaseModel):
+    id: str
+    username: str
+    email: str
+    is_active: bool = True
+
+class UserCreate(BaseModel):
+    username: str
+    email: str
+    password: str
+
+class UserUpdate(BaseModel):
+    username: Optional[str] = None
+    email: Optional[str] = None
+
+class Task(BaseModel):
+    id: str
+    title: str
+    description: Optional[str] = None
+    status: str
+    assigned_to: Optional[str] = None
+
+class TaskCreate(BaseModel):
+    title: str
+    description: str
+    status: str
+    assigned_to: str
+
+class TaskUpdate(BaseModel):
+    title: Optional[str] = None
+    description: Optional[str] = None
+    status: Optional[str] = None
+    assigned_to: Optional[str] = None
+
+class AuthRequest(BaseModel):
+    username: str
+    password: str
+
+class AuthResponse(BaseModel):
+    access_token: str
+    token_type: str = "bearer"
