+++ view/__init__.py
@@ -1,0 +1,1 @@
+from view.users import router as users_router
